PerformanceManager: Don't override the user preference
The current code can override the user preference in different ways.
For instance, entering and exiting the powersaving mode will set
the balanced profile, no matter what was the user preference.

mUserProfile is now updated only when the user changes the profile
and not when the profile is changed automatically (e.g. when
enetering the powersaving mode). When mUserProfile is updated,
the value in the database is also updated, except when initialized.
This also includes the case in which the powersave mode is enabled.

Change-Id: Id88e64389ccb877ebee665d90bf081aa94537542